Move encrypt credentials method to base.
This method will be used by both job creator and credetials
services.
